requirements.txt
keyboard==0.13.5
Pillow==9.2.0
Pillow==10.3.0
PyAutoGUI==0.9.54
PyQt5==5.15.10
PyQt5_sip==12.13.0
pytesseract==0.3.10
pywin32==306
